requirements.txt
adal==1.2.7
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.6
altair==4.1.0
anyio==3.5.0
argcomplete==1.12.3
argon2-cffi==21.1.0
asgiref==3.5.0
astor==0.8.1
async-timeout==4.0.2
attrs==21.2.0
azure-common==1.1.28
azure-core==1.21.1
azure-graphrbac==0.61.1
azure-mgmt-authorization==0.61.0
azure-mgmt-containerregistry==8.2.0
azure-mgmt-core==1.3.0
azure-mgmt-keyvault==9.3.0
azure-mgmt-resource==20.1.0
azure-mgmt-storage==19.1.0
azureml-core==1.39.0
backcall==0.2.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.1.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
blinker==1.4
boto3==1.21.16
botocore==1.24.16
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==7.1.2
cloudpickle==2.0.0
contextlib2==21.6.0
cryptography==36.0.0
cycler==0.11.0
databricks-cli==0.16.4
dataframe-image==0.1.1
debugpy==1.5.1
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
docker==5.0.3
entrypoints==0.3
et-xmlfile==1.1.0
fastapi==0.75.0
Flask==2.0.2
fonttools==4.28.2
fpdf2==2.4.6
frozenlist==1.3.0
gcloud==0.18.3
gitdb==4.0.9
GitPython==3.1.24
google-api-core==2.7.1
google-auth==2.6.0
google-cloud-core==2.2.3
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.53.0
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
h11==0.13.0
htmlmin==0.1.12
httplib2==0.20.2
humanfriendly==10.0
idna==3.3
ImageHash==4.2.1
importlib-metadata==4.11.2
ipykernel==6.5.1
ipython==7.29.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isodate==0.6.1
itsdangerous==2.0.1
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.0.1
jsonpickle==2.1.0
jsonschema==4.2.1
jupyter-client==7.0.6
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jwcrypto==1.0
kiwisolver==1.3.2
knack==0.9.0
kubernetes==23.3.0
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
missingno==0.5.1
mistune==0.8.4
mlflow==1.24.0
mlserver==1.0.1
mlserver-mlflow==1.0.1
moncli==2.0.2
msal==1.17.0
msal-extensions==0.3.1
msrest==0.6.21
msrestazure==0.6.4
multidict==6.0.2
multimethod==1.8
nbclient==0.5.9
nbconvert==6.3.0
nbformat==5.1.3
ndg-httpsclient==0.5.1
nest-asyncio==1.5.1
networkx==2.8
notebook==6.4.6
numpy==1.21.4
oauth2client==4.1.3
oauthlib==3.2.0
openpyxl==3.0.9
packaging==21.3
pandas==1.3.4
pandas-profiling==3.1.0
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.8.2
pathspec==0.9.0
pexpect==4.8.0
phik==0.12.2
pickleshare==0.7.5
Pillow==8.4.0
pkginfo==1.8.2
plotly==5.6.0
portalocker==2.4.0
prometheus-client==0.12.0
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.22
protobuf==3.19.1
ptyprocess==0.7.0
py-grpc-prometheus==0.7.0
pyarrow==6.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.21
pycryptodome==3.11.0
pydantic==1.9.0
pydeck==0.7.1
Pygments==2.10.0
PyJWT==2.3.0
Pympler==0.9
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
Pyrebase4==4.5.0
pyrsistent==0.18.0
pysftp==0.2.9
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==3.3.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
requests==2.26.0
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
rsa==4.7.2
s3transfer==0.5.2
schematics==2.1.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
SecretStorage==3.3.1
semver==2.13.0
Send2Trash==1.8.0
setuptools-scm==6.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
starlette==0.17.1
starlette-exporter==0.12.0
streamlit==1.7.0
streamlit-pandas-profiling==0.1.3
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.1.0
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
uvicorn==0.17.5
validators==0.18.2
visions==0.7.4
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.2
widgetsnbextension==3.5.2
wrapt==1.13.3
yarl==1.7.2
zipp==3.7.0
